feat(filter-tag): refactor
