+++ packages/filter-tag/src/helpers/is-keyboard-event.ts
@@ -1,0 +1,6 @@
+import { KeyboardEvent, MouseEvent } from 'react';
+
+export const isKeyBoardEvent = (
+    event: MouseEvent<HTMLDivElement> | KeyboardEvent<HTMLDivElement>,
+): event is KeyboardEvent<HTMLDivElement> =>
+    (event as KeyboardEvent<HTMLDivElement>).key !== undefined;
