+++ packages/filter-tag/src/mobile/Component.mobile.tsx
@@ -1,6 +1,8 @@
 import React, { forwardRef } from 'react';
 
-import { BaseFilterTag, BaseFilterTagProps, PrivateProps } from '../components/base-filter-tag';
+import { BaseFilterTag } from '../components/base-filter-tag';
+import { PrivateProps } from '../types/base-filter-tag-private-props';
+import { BaseFilterTagProps } from '../types/base-filter-tag-props';
 
 import defaultColors from './default.mobile.module.css';
 import invertedColors from './inverted.mobile.module.css';
